runtime: atomically load and update panicking
This closes a minor race where we could print two racing panics via
winthrow.

That said, I believe the primary purpose of this check is to prevent
infinite recursion of exceptions during exception handling on a single
thread. For that purpose, an even beeter follow-up will be to use the
gp.m.dying escalation in startpanic_m.

Change-Id: Icdd5dff1831d0cfa85ea5c313714f38689409aa8