runtime: replace custom dying check with startpanic_m
winthrow uses the panicking counter as a basic check for nested
exceptions, used to avoid infinite recursion of exceptions if we crash
while trying to to die.

startpanic_m performs that same kind of check with gp.m.dying.
startpanic_m's check is a bit more robust by allowing a few levels of
recursion with informational messages along the way to make it more
clear what is happening.

Use this for winthrow to get the advantages there.

Change-Id: If7c5c68425ca29275798d0225f6bc6a2812ec00c