cmd/compile: skip doing comparison for invalid types
CL 381964 made a better error message for comparisons, but relying on
the "hasNil" check for reporting error in case of non-pointers compare
with nil. For "*T" where "T" is invalid type, we return "Typ[Invalid]"
instead of "*T", to terminate the typecheck process as soon as possible,
and not show some confused errors (CL 356449). It makes the "hasNil"
returning wrong result for "*T", causing confused follow-on errors.

To fix this, just skipping the comparison check if type is invalid.

Fixes #54405

Change-Id: Ic219f8fe20d8e56d8740570928dda79cdf1f8d59