go/types: optimize lookupFieldOrMethod
LookupFieldOrMethod appeared as a hotspot when benchmarking gopls'
auto-completion. In particular, instanceLookup.add was allocating in the
common case of structs with no embedding.

This is easily fixed, by using a small array as a front-buffer for the
lookup. Do this, and add a benchmark.

WIP: need to ensure this doesn't affect validType.
DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I10b6deb3819ab562dbbe1913f12b977cf956dd50