cmd/dist: enable testsanitizer test on ppc64le
Even though the -race option works for ppc64le, some tests in the
testsanitizer test fail on the builders, even though they don't fail
on any of our systems.

This is an experiment to re-enable this test on ppc64le to see if
it still fails on the builders.

Updates #45040

Change-Id: I0729bec5864e6c4cc752968485e89179df027063