internal/sysinfo: consolidate kernelVersion implementations
Currently, there are 3 functions returning Linux kernel version numbers.
Two of them (in internal/poll and net) are identical, and the third is
a more complex regexp-based one (in runtime/pprof) used for a test.

Remove the complex implementation, and move the simple one into
internal/sysinfo. Use it from all the three places.

Change-Id: I4a34d9ca47257743c16def30e4dd634e36056091