go/internal/gcimporter: further simplify unified IR importer
Now that we don't delay calling SetUnderlying (CL 424876), there
doesn't appear to be any need to delay calling SetConstraint either.

However, I'm a bit iffier on why this was originally needed and
whether there may be cases where it's still needed. Hence splitting it
out separately from the previous CL, to make it easier to revert if
necessary.

Change-Id: Ib426079b13e8a2dc17796691b6e049dde1e681c1