go/token: make mutex locking in unpack cheaper
I was profiling the cpu usage of go/printer's only benchmark,
and found that token.File.Unpack was one of the top offenders.

It was mainly the deferred unlock that took a big chunk of time,
and to my surprise, reoving the use of defer helped significantly:

	name      old time/op    new time/op    delta
	Print-16    5.61ms ± 2%    5.38ms ± 1%  -4.04%  (p=0.000 n=10+8)

	name      old speed      new speed      delta
	Print-16  9.27MB/s ± 2%  9.64MB/s ± 1%  +4.03%  (p=0.000 n=9+8)

	name      old alloc/op   new alloc/op   delta
	Print-16     332kB ± 0%     332kB ± 0%    ~     (p=0.363 n=10+10)

	name      old allocs/op  new allocs/op  delta
	Print-16     3.45k ± 0%     3.45k ± 0%    ~     (all equal)

DO NOT SUBMIT: Why is this the case? Surely a single defer shouldn't
incur such a large cost? Could it be that the naked return is somehow
related? Perhaps the compiler people can take a look.

Note that I re-ran the benchmarks multiple times, and a 4% speed-up is
too significant to simply be a coincidence.

Change-Id: I5a1c6d36a8e8357435a305a1bc0970ee0358b08a