go/printer: make space for the output buffer
First, we know that Go source files almost always weigh at least a few
kilobytes, so we can kickstart the output buffer to be a reasonable size
and reduce the initial number of incremental allocations and copies when
appending bytes or strings to output.

Second, in nodeSize we use a nested printer, but we don't actually need
its printed bytes - we only need to know how many bytes it prints.
For that reason, use a throwaway buffer: the part of our output buffer
between length and capacity, as we haven't used it yet.

Both of those combined result in a modest reduction in allocations and
CPU usage. The benchmark uses testdata/parser.go, which has just over
two thousand lines of code, which is pretty standard size-wise.

We also split the Print benchmark to cover both a medium-sized ast.File,
as well as a pretty small ast.Decl node. The latter is a somewhat common
scenario in gopls, which has code actions which alter small bits of the
AST and print them back out to rewrite only a few lines in a file.

	name          old time/op    new time/op     delta
	PrintFile-16    5.43ms ± 1%     5.18ms ± 1%    -4.64%  (p=0.000 n=9+9)
	PrintDecl-16    19.1µs ± 0%     24.8µs ± 1%   +29.45%  (p=0.000 n=10+10)

	name          old speed      new speed       delta
	PrintFile-16  9.56MB/s ± 1%  10.01MB/s ± 1%    +4.71%  (p=0.000 n=8+9)
	PrintDecl-16  1.67MB/s ± 0%   1.29MB/s ± 1%   -22.76%  (p=0.000 n=10+10)

	name          old alloc/op   new alloc/op    delta
	PrintFile-16     332kB ± 0%      276kB ± 0%   -16.92%  (p=0.000 n=10+8)
	PrintDecl-16    3.92kB ± 0%    19.70kB ± 0%  +402.45%  (p=0.000 n=10+10)

	name          old allocs/op  new allocs/op   delta
	PrintFile-16     3.45k ± 0%      2.83k ± 0%   -17.98%  (p=0.000 n=10+10)
	PrintDecl-16      56.0 ± 0%       50.0 ± 0%   -10.71%  (p=0.000 n=10+10)

Change-Id: I475a3babca77532b2d51888f49710f74763d81d2