go/printer: make space for the output buffer
First, we know that Go source files almost always weigh at least a few
kilobytes, so we can kickstart the output buffer to be a reasonable size
and reduce the initial number of incremental allocations and copies when
appending bytes or strings to output.

Second, in nodeSize we use a nested printer, but we don't actually need
its printed bytes - we only need to know how many bytes it prints.
For that reason, use a throwaway buffer: the part of our output buffer
between length and capacity, as we haven't used it yet.

Both of those combined result in a modest reduction in allocations and
CPU usage. The benchmark uses testdata/parser.go, which has just over
two thousand lines of code, which is pretty standard size-wise.

	name      old time/op    new time/op    delta
	Print-16    5.62ms ± 2%    5.43ms ± 1%   -3.34%  (p=0.000 n=10+10)

	name      old speed      new speed      delta
	Print-16  9.24MB/s ± 2%  9.55MB/s ± 1%   +3.43%  (p=0.000 n=10+10)

	name      old alloc/op   new alloc/op   delta
	Print-16     332kB ± 0%     276kB ± 0%  -16.93%  (p=0.000 n=10+10)

	name      old allocs/op  new allocs/op  delta
	Print-16     3.45k ± 0%     2.83k ± 0%  -17.98%  (p=0.000 n=10+10)

Change-Id: I475a3babca77532b2d51888f49710f74763d81d2