net/netip: fix invalid representation of Prefix
For a given Addr, make it such that there is exactly one invalid representation.
This allows invalid representations to also be comparable.
To ensure that the zero value of Prefix is invalid, we modify the encoding
of bits to simply be the bit count plus one.

Since Addr is immutable, we check in the PrefixFrom constructor that
the provided Addr is valid and only store a non-zero bits length if so.
This simplifies IsValid to simply checking whether bitsPlusOne is non-zero.

Fixes #54525

Change-Id: I9244cae2fd160cc9c81d007866992df2e422d3b9