runtime: use SwissTable
We suggest using SwissTable in the runtime to replace the original implementation of the hashmap.

SwissTable is now used in abseil and rust by default. We borrowed ideas from the two implementations and combined some optimization ideas of Go's original hashmap to construct the current implementation.

- abseil: https://abseil.io/about/design/swisstables
- rust: https://github.com/rust-lang/hashbrown

map properties changed with this CL:
- maps immediately grow in place leader to potential long wait on map assign
- It might require more memory for some maps (there can be cases where before we had spare overflow buckets and now we don't use them, also 7 items max per bucket not 8)

For #54766

Change-Id: I66f7de17d8b36b578f9cbc71843b0e4356b914f2