cmd/link:  support LLVM/Clang for MSVC
I find that cgo can compatible with LLVM/Clang for MSVC well actually.
The only problem I have found is that some arguments for linker of Mingw/GCC can't be used in linker of MSVC, so I write a judgement.
If go detect that the toolchain is LLVM/Clang for MSVC, do not pass those arguments that are not supported by linker of  MSVC.

Change-Id: Ib4e9507755ce9f91c020c15c03375fc56753644b
GitHub-Last-Rev: a94ddccffa3ec0747c9550f21054ed8683741458
GitHub-Pull-Request: golang/go#54810