cmd/link: support LLVM/Clang for MSVC
I find that cgo can compatible with LLVM/Clang for MSVC well actually.
The only problem I have found is that some arguments for linker of Mingw/GCC can't be used in linker of MSVC, so I write a judgement.
If go detect that the toolchain is LLVM/Clang for MSVC, do not pass those arguments that are not supported by linker of MSVC.

Change-Id: I4849087a6e614ff52776305e6755d0b515316de8
GitHub-Last-Rev: 364c5796934fbd35900c2a81b39107edc5963674
GitHub-Pull-Request: golang/go#54811