os: avoid calling copy_file_range(2) when dst and src are the same file
If fd_in and fd_out refer to the same file and the source and target ranges overlap,
copy_file_range(2) will just return EINVAL error.

We can't tell if two different file descriptors pointing to the same file overlap,
but if these two are the same fd, they must overlap, because we call copy_file_range(2)
with off_in==NULL and off_out==NULL, so don't bother to try and avoid this system call.

Change-Id: Ib9746cb4f27625cb22620271b280d2da242b2fba