os: fix wrong error msg from TestDoubleCloseError
When the type assertion fails, the test mistakenly prints the expected
(rather than the actual) type.

When the error string doesn't match, the text mistakenly prints the
original (rather than the converted) error (although there might not be
any difference in the output, the code looks wrong).

Fix both issues.

Fixes: e3d7ec006f2538597 ("os: consistently return ErrClosed for closed file")
Change-Id: Ia7dd0632fc677f458fec25d899c46268a12f76e7