internal/testenv: add AugmentToolBuildForCoverage
Introduce a new testenv helper AugmentToolBuildForCoverage intended to
be used by tests in the "cmd" subdir that build copies of themselves,
to help with collecting code coverage.

If a unit test for cmd/foo builds a copy of the cmd/foo executable and
then runs tests against it, then when someone comes along and runs "go
test -cover cmd/foo", we want the test that rebuilds cmd/foo to build
that new executable with -cover as well. AugmentToolBuildForCoverage
automates/simplifies this process.

Change-Id: I9abeb78062de8ffda73cdcb437cf4f8299347a22