runtime: remove atomic store requirement on pageAlloc.chunks
pageAlloc.chunks used to require an atomic store when growing the heap
because the scavenger would look at the list without locking the heap
lock. However, the scavenger doesn't do that anymore, and it looks like
nothing really does at all.

We can't actually make the store non-atomic because that invokes a write
barrier (not safe here) but we can at least update the documentation.

Change-Id: Ib452d147861060f9f6e74e2d98ee111cf89ce8f5